+++ src/Domain/ProductVariants/Http/Routing/ProductVariantRouteGroup.php
@@ -31,6 +31,7 @@ class ProductVariantRouteGroup extends RouteGroup implements RouteGroupContract
                         $relationships->hasMany('product')->readOnly();
                         $relationships->hasOne('thumbnail')->readOnly();
                         $relationships->hasMany('urls')->readOnly();
+                        $relationships->hasMany('product_option_values')->readOnly();
                     })
                     ->only('index', 'show')
                     ->readOnly();